Becomes able to handle changed format in eventfile
Earlier the solution data had to be given in each Solution in event file.
But now all the solution data can be kept at one place and the whole
process becomes pretty neat. For this it was necessary to make changes to
the way data is read in master_controller.pl.